refactor(blorktools): delete glb manager
