+++ packages/blorktools/src/asset_debugger/util/scene/glb-controller.js
@@ -1,7 +1,7 @@
 import { processGLBFile } from '../data/upload/glb-file-handler.js';
 import { getState, updateState } from '../state/scene-state.js';
-import { getCurrentGlbBuffer, setCurrentGlbBuffer } from './glb-manager.js';
 
+let currentGlbBuffer = null;
 let bufferUpdateListeners = [];
 let initialized = false;
 
@@ -177,4 +177,32 @@ export async function downloadUpdatedGlb() {
 
 function getCurrentTimestamp() {
     return new Date().toISOString().replace(/[:.]/g, '-').slice(0, -5);
+}
+
+export function getMeshByIndex(meshIndex) {
+    const state = getState();
+    if (!state || !state.meshes || meshIndex >= state.meshes.length) {
+        return null;
+    }
+    
+    return state.meshes[meshIndex];
+}
+
+
+export function setCurrentGlbBuffer(glbBuffer) {
+    currentGlbBuffer = glbBuffer;
+}
+
+export function getCurrentGlbBuffer() {
+    if (currentGlbBuffer) {
+        return currentGlbBuffer;
+    }
+    
+    const state = getState();
+    if (state && state.currentGlb && state.currentGlb.arrayBuffer) {
+        currentGlbBuffer = state.currentGlb.arrayBuffer;
+        return currentGlbBuffer;
+    }
+    
+    return null;
 }
