+++ packages/blorktools/src/asset_debugger/util/scene/threejs-model-manager.js
@@ -10,8 +10,7 @@ import { getState, updateState } from '../state/scene-state.js';
 import { createMaterial } from '../animation/render/pbr-material-factory.js';
 import { fitCameraToObject } from '../../util/scene/threejs-scene-controller.js';
 import { updateUvPanel } from '../../panels/asset-panel/uv-heading/uv-heading.js';
-import { getCurrentGlbBuffer, setCurrentGlbBuffer } from './glb-manager.js';
-import { processModelFileForHtmlEditor } from './glb-controller.js';
+import { processModelFileForHtmlEditor, getCurrentGlbBuffer, setCurrentGlbBuffer } from './glb-controller.js';
 import { createMeshVisibilityPanel } from '../../panels/asset-panel/mesh-heading/mesh-heading.js';
 import { updateAtlasVisualization } from '../../panels/asset-panel/atlas-heading/atlas-heading.js';
 import { updateRigPanel } from '../../panels/asset-panel/rig-heading/rig-heading.js';
